requirements.txt
aniso8601==1.2.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
feedparser==5.2.1
Flask==0.12.2
Flask-RESTful==0.3.6
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
wheel==0.24.0
xmltodict==0.11.0
